Default volume change, I2C begin restricted, more randomly chosen sounds added
Default volume set to 5.0 as this does not scale the audio in software. Allows selection of 3.0 to 5.0 in 1.0 steps.

changed called to DEC_I2C.begin() to only take place if either or both enableLightSensor or enableTimeOfFlightSensor is true.

Added additional single char commands for playing individual sounds on demand at random. Allows for multiple sets of guidance sounds to be defined and commanded.